refactor(orders): extract pickup module + enforce picked_up CHECK
- Move createOrderPickupEvent + pickup presign into order-pickup.service.ts
- Replace hand-rolled compensating rollback with a db.transaction
- Add CHECK order_services_picked_up_requires_event_check (ADR-0005 fix)
- Repoint route imports; drop orphaned imports in order-admin.service.ts
